upload venv zip file
venv zip file is a virtual environment that I established on PyCharm, if you use the PyCharm but you do not know how to configure the virtualenv environment, please try to download the venv zip and put it on the root directory same as mapp folder or something else. Try it out!